Add dark theme for kitty + scripts
This commit adds a dark theme for kitty.
Using the scripts under `scripts/`, the kitty theme can be set
accordingly. Also the `~/.darktheme` file will be made/removed for vim's
theme. When changing the theme, kitty can be reloaded with
<Ctrl>+<Shift>+<F5> to apply the new config. vim's theme can still be
changed with `:ToggleDark`.
Also the setup scripts are now made silent, if they worked. For the
tmux, vim and kitty scripts, there is now a check, if the corresponding
program is installed on the system. If not, an error gets displayed and
the script does not run further actions.